+++ generator/datanode/templates.go
@@ -16,8 +16,12 @@ import (
 	"github.com/BurntSushi/toml"
 )
 
+type IPSFPeer struct {
+	Index int
+	ID    string
+}
+
 type ConfigTemplateContext struct {
-	Prefix      string
 	NodeHomeDir string
 	NodeNumber  int
 	NodeSet     types.NodeSet
@@ -47,8 +51,8 @@ func (tc ConfigTemplateContext) GetDehistoryPrivKey(nodeNumber int) string {
 	return id.PrivKey
 }
 
-func (tc ConfigTemplateContext) IPSFPeersIDs() []string {
-	peersIDs := []string{}
+func (tc ConfigTemplateContext) IPSFPeers() []IPSFPeer {
+	peersIDs := []IPSFPeer{}
 	for _, node := range tc.nodes {
 		if len(tc.nodes) != 1 && node.index == tc.NodeSet.Index {
 			continue
@@ -60,7 +64,10 @@ func (tc ConfigTemplateContext) IPSFPeersIDs() []string {
 			panic("couldn't create ipfs peer identity")
 		}
 
-		peersIDs = append(peersIDs, id.PeerID)
+		peersIDs = append(peersIDs, IPSFPeer{
+			Index: node.index,
+			ID:    id.PeerID,
+		})
 	}
 
 	return peersIDs
@@ -77,7 +84,6 @@ func NewConfigTemplate(templateRaw string) (*template.Template, error) {
 
 func (dng ConfigGenerator) TemplateConfig(ns types.NodeSet, configTemplate *template.Template) (*bytes.Buffer, error) {
 	templateCtx := ConfigTemplateContext{
-		Prefix:      *dng.conf.Prefix,
 		NodeNumber:  ns.Index,
 		NodeHomeDir: dng.homeDir,
 		NodeSet:     ns,